requirements.txt
anyio==4.4.0
arrow==1.3.0
asttokens==2.4.1
attrs==24.2.0
binaryornot==0.4.4
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.53.1
h11==0.14.0
humanize==4.10.0
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
ipyvue==1.11.1
ipyvuetify==1.10.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
Mesa==2.3.4
Mesa-Viz-Tornado==0.1.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
pymdown-extensions==10.9
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
reacton==1.8.3
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rich-click==1.8.3
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
solara==1.39.0
solara-server==1.39.0
solara-ui==1.39.0
stack-data==0.6.3
starlette==0.38.5
text-unidecode==1.3
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
watchdog==5.0.2
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
widgetsnbextension==4.0.13
